% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_preproc.R
\name{eq_CleanData}
\alias{eq_CleanData}
\title{clean/format existing variables for earthquake data.}
\usage{
eq_CleanData(df)
}
\arguments{
\item{df}{data.frame - earthquake data}
}
\value{
data.frame - cleaned earthquake data
}
\description{
Moreover it generates additinal variables - MAGNITUDE and DATE which is important at vizualization step.
}
\examples{

df = eq_ReadQuake()
eq_CleanData(df)

}
\seealso{
\code{\link{eq_ReadQuake}} \code{\link{eq_GenLocationName}} \code{\link{geom_TimeLine}} \code{\link{geom_TimeLineLable}}
}
