% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_CreateLabel}
\alias{eq_CreateLabel}
\title{Supporting function for \code{eq_Map}. It should be used for genereting more informative labels.}
\usage{
eq_CreateLabel(df)
}
\arguments{
\item{df}{data.frame - cleaned earthquake data}
}
\value{
data.frame - earthquake data with an addditional event description variable in a html format.
}
\examples{

df = eq_ReadQuake()

library(dplyr)

eq_CleanData(df) \%>\%
eq_GenLocationName() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 1980) \%>\%
eq_CreateLabel() \%>\%
eq_Map(label = "popup_text")

}
\seealso{
\code{\link{eq_Map}}
}
