% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_preproc.R
\name{eq_GenLocationName}
\alias{eq_GenLocationName}
\title{Function is provided to process LOCATION_NAME variable in such way to subset only a region name.}
\usage{
eq_GenLocationName(df, to_name = "LOCATION")
}
\arguments{
\item{df}{character - a path to the file or the file name}

\item{to_name}{character - a path to the file or the file name}
}
\value{
data.frame - data expanded by LOCATION variable
}
\description{
If event occur at few locations the function return region names for all places.
}
\examples{

 df = eq_ReadQuake()
 eq_GenLocationName(eq_CleanData(df))

}
\seealso{
\code{\link{eq_ReadQuake}} \code{\link{eq_ReadQuake}} \code{\link{geom_TimeLine}} \code{\link{geom_TimeLineLable}}
}
