% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_preproc.R
\name{eq_GetPosixDate}
\alias{eq_GetPosixDate}
\title{Component of \code{eq_CleanData} - genereting POSIX Time variable}
\description{
Component of \code{eq_CleanData} - genereting POSIX Time variable
}
