% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_Map}
\alias{eq_Map}
\title{Function for plotting a leaflet map of earthquake occurences.}
\usage{
eq_Map(df, label = "DATE")
}
\arguments{
\item{df}{data.frame - cleaned earthquake data}

\item{label}{character - Column in the tbl_df object to be used for annotation.}
}
\value{
leaflet map - interactive map of significant earthquakes
}
\examples{

df = eq_ReadQuake()

library(dplyr)

eq_CleanData(df) \%>\%
eq_GenLocationName() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 1980) \%>\%
eq_Map(label = "DATE")

}
\references{
\url{http://rstudio.github.io/leaflet/}
}
\seealso{
\code{\link{eq_CreateLabel}}
}
