% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_preproc.R
\name{eq_ReadQuake}
\alias{eq_ReadQuake}
\title{Checks if the file exists and then load it as data.frame.}
\source{
National Geophysical Data Center / World Data Service NGDC/WDS: Significant Earthquake Database. National Geophysical Data Center, NOAA. doi:10.7289/V5TD9V7K
}
\usage{
eq_ReadQuake(filename = "")
}
\arguments{
\item{filename}{character - a path to the file or the file name. Default value providing pre-loaded database.}
}
\value{
data in a data.frame format
}
\examples{

eq_ReadQuake()

}
\seealso{
\code{\link{eq_CleanData}} \code{\link{eq_GenLocationName}} \code{\link{geom_TimeLine}} \code{\link{geom_TimeLineLable}}
}
